Restructure some more and get tests working
