+++ predicate/pyproject.toml
@@ -13,6 +13,7 @@ flake8 = "^5.0.4"
 isort = "^5.10.1"
 pytest = "^7.1.3"
 black = "^22.8.0"
+setuptools = "^65.3.0"
 
 [tool.lint-python]
 lint-version = "2"
