+++ predicate/src/solver/__init__.py
