+++ predicate/src/solver/errors.py
