+++ predicate/src/solver/test/__init__.py
@@ -1,0 +1,1 @@
+# Tests and examples
